WIP: PostgreSQL with Sequelize for a GraphQL Server is next
