+++ api/src/resolvers/index.js
@@ -1,63 +1,4 @@
-const uuidv4 = require('uuid/v4')
+const userResolvers = require('./user')
+const messageResolvers = require('./message')
 
-module.exports = {
-  Query: {
-    users: (parent, args, { models }) => {
-      return Object.values(models.users);
-    },
-    user: (parent, { id }, { models }) => {
-      return models.users[id];
-    },
-    me: (parent, args, { me }) => {
-      return me;
-    },
-    messages: (parent, args, { models }) => {
-      return Object.values(models.messages);
-    },
-    message: (parent, { id }, { models }) => {
-      return models.messages[id];
-    },
-  },
-
-  Mutation: {
-    createMessage: (parent, { text }, { me, models }) => {
-      const id = uuidv4();
-      const message = {
-        id,
-        text,
-        userId: me.id,
-      };
-
-      models.messages[id] = message;
-      models.users[me.id].messageIds.push(id);
-
-      return message;
-    },
-
-    deleteMessage: (parent, { id }, { models }) => {
-      const { [id]: message, ...otherMessages } = models.messages;
-
-      if (!message) {
-        return false;
-      }
-
-      models.messages = otherMessages;
-
-      return true;
-    },
-  },
-
-  User: {
-    messages: (user, args, { models }) => {
-      return Object.values(models.messages).filter(
-        message => message.userId === user.id,
-      );
-    },
-  },
-
-  Message: {
-    user: (message, args, { models }) => {
-      return models.users[message.userId];
-    },
-  },
-};
+module.exports = [userResolvers, messageResolvers];
