+++ api/src/schema/user.js
@@ -1,4 +1,4 @@
-import { gql } from 'apollo-server-express';
+const { gql } = require('apollo-server-express')
 
 module.exports = gql`
   extend type Query {
